+++ tools/perf/util/intlist.h
@@ -9,7 +9,7 @@
 
 struct int_node {
 	struct rb_node rb_node;
-	unsigned long i;
+	int i;
 	void *priv;
 };
 
@@ -21,13 +21,13 @@ struct intlist *intlist__new(const char *slist);
 void intlist__delete(struct intlist *ilist);
 
 void intlist__remove(struct intlist *ilist, struct int_node *in);
-int intlist__add(struct intlist *ilist, unsigned long i);
+int intlist__add(struct intlist *ilist, int i);
 
 struct int_node *intlist__entry(const struct intlist *ilist, unsigned int idx);
-struct int_node *intlist__find(struct intlist *ilist, unsigned long i);
-struct int_node *intlist__findnew(struct intlist *ilist, unsigned long i);
+struct int_node *intlist__find(struct intlist *ilist, int i);
+struct int_node *intlist__findnew(struct intlist *ilist, int i);
 
-static inline bool intlist__has_entry(struct intlist *ilist, unsigned long i)
+static inline bool intlist__has_entry(struct intlist *ilist, int i)
 {
 	return intlist__find(ilist, i) != NULL;
 }
